pyproject.toml
[project]
name = "proj1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "uvicorn>=0.37.0",
]
